Last commit from old sys
Check before merging into default branch

+++ src/main/java/org/parryapplications/spring/todoproject/service/TodoServiceImpl.java
@@ -1,5 +1,6 @@
 package org.parryapplications.spring.todoproject.service;
 
+import jakarta.transaction.Transactional;
 import org.parryapplications.spring.todoproject.dto.TodoDto;
 import org.parryapplications.spring.todoproject.model.Todo;
 import org.parryapplications.spring.todoproject.repository.TodoJpaRepository;
@@ -10,22 +11,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ObjectUtils;
 
-import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Predicate;
 
 @Service
 public class TodoServiceImpl {
 
-    Logger logger = LoggerFactory.getLogger(TodoServiceImpl.class);
-
     private final TodoJpaRepository todoRepository;
     private final CommonUtil commonUtil;
+    Logger logger = LoggerFactory.getLogger(TodoServiceImpl.class);
 
     @Autowired
-    public TodoServiceImpl(TodoJpaRepository todoRepository, CommonUtil commonUtil){
+    public TodoServiceImpl(TodoJpaRepository todoRepository, CommonUtil commonUtil) {
         this.todoRepository = todoRepository;
         this.commonUtil = commonUtil;
     }
@@ -43,38 +40,39 @@ public class TodoServiceImpl {
             Todo todoModel = todoRepository.save(commonUtil.convertTodoDtoToTodoModel(todoDto));
             logger.info("Todo created successfully");
             return commonUtil.convertTodoModelToTodoDto(todoModel);
-        } catch(Exception e){
-            logger.error("Error while creating todo : ",e);
+        } catch (Exception e) {
+            logger.error("Error while creating todo : ", e);
         }
         return null;
     }
 
-    public TodoDto getTodoById(Integer id){
+    public TodoDto getTodoById(String username, Integer id) {
         Optional<Todo> opTodo = todoRepository.findById(id);
-        if(opTodo.isPresent()) {
+        if (opTodo.isPresent() && opTodo.get().getUsername().equals(username)) {
             logger.info("Retrieved Todo By Id");
             return commonUtil.convertTodoModelToTodoDto(opTodo.get());
         }
 
-        logger.error("Error while retrieving todo by id : "+id);
+        logger.info("No todo with ID: " + id + " found from your Data");
         return null;
 
         //        Predicate<? super Todo> predicate = todo -> todo.getId().equals(id);
 //        return todoList.stream().filter(predicate).findAny().orElse(null);
     }
 
-    public List<TodoDto> getAllTodos(){
+    public List<TodoDto> getAllTodos(String username) {
         logger.info("Retrieving all Todos");
-        List<TodoDto> list = todoRepository.findAll().stream().map(commonUtil::convertTodoModelToTodoDto).toList();
+        List<TodoDto> list = todoRepository.findByUsername(username).stream().map(commonUtil::convertTodoModelToTodoDto).toList();
         return list;
     }
 
-    public String deleteTodoById(Integer id){
+    public String deleteTodoById(String username, Integer id) {
         try {
-            todoRepository.deleteById(id);
+            if(getTodoById(username, id) != null)
+                todoRepository.deleteById(id);
 //            todoList.removeIf((todo) -> todo.getId().equals(id));
-        } catch(Exception e){
-            logger.error("Error while delete todo id : {} :: {} ",id,e.getMessage());
+        } catch (Exception e) {
+            logger.error("Error while delete todo id : {} :: {} ", id, e.getMessage());
             return "Error while deleting todo id : " + id;
         }
         logger.info("Todo deleted successfully");
@@ -83,7 +81,7 @@ public class TodoServiceImpl {
 
     public TodoDto updateTodo(TodoDto todoDto) {
         try {
-            TodoDto todoModelById = getTodoById(todoDto.getId());
+            TodoDto todoModelById = getTodoById(todoDto.getUsername(), todoDto.getId());
 
             //Checking if this todo is existed for modify operation:
             if (!ObjectUtils.isEmpty(todoModelById)) {
@@ -95,21 +93,34 @@ public class TodoServiceImpl {
 //            todoList.add(todo);
 //            return todo;
             }
-        } catch(Exception e){
-            logger.error("Error updating todo : {} ",e.getMessage());
+        } catch (Exception e) {
+            logger.error("Error updating todo : {} ", e.getMessage());
             return null;
         }
         return null;
     }
 
-    public String deleteAllTodos() {
+    @Transactional
+    public String deleteAllTodosByUsername(String username) {
         try {
+            int count = todoRepository.deleteAllByUsername(username);
+            logger.info(count + " - Counts have been deleted");
+        } catch (Exception e) {
+            logger.error("Error while deleting all todos by username :: ",e);
+            return "Error while deleting all todos by username";
+        }
+        logger.info("All todos as per username have been deleted successfully");
+        return "All " +username+ " Todos Deleted Successfully";
+    }
+
+    @Transactional
+    public String deleteAllTodos() {
+        try{
             todoRepository.deleteAll();
-        } catch(Exception e){
-            logger.error("Error while deleting all todos");
+        } catch (Exception e){
+            logger.error("Error while deleting all todos :: ",e);
             return "Error while deleting all todos";
         }
-        logger.info("All todos deleted successfully");
-        return "All Todos Deleted Successfully";
+        return "All Todos deleted successfully";
     }
 }
